Reporting changes (#7646)
Reintroduce corrected #7383

This switches arcade from using a custom reporter, to using a reporter more tightly integrated with helix. That reporter has the capacity to handle local test retrying and reporting to Azure DevOps for "flaky" tests.

To enable the retry, a file "eng/test-configuration.json" must be created following the documented schema.