Skip validating submodules sourcelinks
Currently, sourcelink is broken for submodules, where it will assume that the link is actually in the current repo, rather than having info based on the actual repo the submodule came from. There is work to fix submodule links in sourcelink, but until that fix is in, we should just ignore all links are from submodules.