Remove Bundle override of DOTNETHOME (#7855)
The bundle was passing in a hardcoded DOTNETHOME value unconditionally
to MSIs in the chain.  This value doesn't have the redirection we're
adding to the MSI.  We can't add that redirection in the bundle, since
Burn doesn't have conditional variables.  It can only set variables as a
result of a search to either the value of the search (if successful) or
1/0.