Support Apple app bundle signing for already archived payloads (#7864)
When user provides an already archived application (or more), we need to inject the provisioning profile in each of the app bundles inside of the archive. This was not supported.

If there are no app bundles (might be built as part of the Helix work item), we only store the profile in the root of the payload.