Add back the quotes to sdl configuration (#8074)
When the code for running the configure step in sdl validation was moved, the quotes were removed from the tool args. This prevents us from upgrading the version of guardian, because newer versions of guardian get confused by the triangle brackets. This change adds back the quotes around each tool arg.