Use XHarness' diagnostics output to detect infrastructure issues (#8139)
Instead of relying on the work item to request retry/reboot, we do it in the SDK at the end of the job while reading the `diagnostics.json` file generated by XHarness.
This will transparently work for any number of XHarness commands executed during the work item and can be a single place for the logic. It also unifies the logic we had duplicated over the `.ps` and `.sh` files.
The user then doesn't need to categorize or react on failed XHarness commands in any way.

We still allow the users to request retry/reboot using the .dotfiles but it should not be necessary.