Add the ability to cache APT packages in build-rootfs.sh
When creating a cross-building rootfs for a debian-based system, the script
build-rootfs.sh downloads packages using `deboo# Problem

When creating a cross-building rootfs for a Debian-based system, the script
`build-rootfs.sh` downloads packages using `debootstrap` and `apt-get`. If
this script is run multiple times, the packages will be downloaded multiple
times. If the network path to the package servers is slow, the build process
can take a while. There is an opportunity to speed up the rootfs building
process by caching the packages locally.

# Solution

This change add the ability for the user to specify a caching server.

The cache server ( [apt-cache-ng](https://wiki.debian.org/AptCacherNg) ) can be installed by running

    apt-get install apt-cacher-ng

The caching server can be used like this:

    ./eng/common/cross/build-rootfs.sh arm64 --apt-proxy http://127.0.0.1:3142

# Benchmarking

My bench marking setup is kind of a worse case for this change: a 10-year old CPU and 1Gbps internet connection.
So the build process downloads the packages fairly fast and installing them takes a while.
Even under these conditions I can easily see a speed up when performing the following commands:

```
apt install apt-cacher-ng
time ./eng/common/cross/build-rootfs.sh arm64 --apt-proxy http://127.0.0.1:3142
git clean -dxf
time ./eng/common/cross/build-rootfs.sh arm64 --apt-proxy http://127.0.0.1:3142
```

The first build took 515 seconds, the second took 474 seconds. That's a 41 second improvement.tstrap` and `apt-get`. If
this script is run multiple times, the packages will be downloaded multiple
times. This can be slow if the network path to the package server is slow

This change add the ability for the user to specify a caching server.

The cache server can be installed by running
    apt-get install apt-cacher-ng

The caching server can be used like this:
    ./eng/common/cross/build-rootfs.sh arm64 --apt-proxy http://127.0.0.1:3142