Use the source directory as the working directory for running sdl (#8167)
We were using the "working directory," which is one level up from the SourceDirectory passed in. This meant that we were unintentionally getting the .gdn directory and anything else that may be laying around on the machine as part of our working directory. In the Validate-DotNet pipeline, it meant we were scanning more than we were meaning to scan.

This change adjusts our call to run-sdl to use the SourceDirectory (ie what we actually WANT to scan) as the working directory.