Allow Publishing to create alternate aka.ms links for installer/sdk
We want to be able to not only specify -Channel 7.0.1xx to install the sdk from daily links, but also -Channel 7.0. In order to do this, we need to be able to create multiple aka.ms links for repos that publish to the SDK channels.

This change adds the "AlternateAkaMSChannelName" property to TargetChannelConfig and "AlternateShortUrlPrefix" to TargetFeedConfig. We then use this alternate prefix, when it exists, to create an additional aka.ms link in the LatestLinksManager.