Fix regression in CoreClrConfigurationDetection
In 16988fc6f9799e2a119167b1730953778e2b0b53 the refactoring added `IsGCStress` and used it in the `IsStressTest` property.
However the `GetEnvironmentVariableValue("GCStress") != string.Empty` is always true since `GetEnvironmentVariableValue` will return the string "0" if the env variable isn't set.

This in turn caused `StressModeApplies` to return false for `RuntimeTestModes.RegularRun` which made the `SkipOnCoreClr` attribute a no-op in regular runs.