Fix aka.ms links when no ChannelNames given (#8236)
When there are no AkaMSChannelNames, we used to default to using dotnet/ as the short link url. When we switched to multiple possible channel names, we removed that default, so any channel with nothing in the list would just not create aka.ms links. This change returns to the previous behavior to use the default dotnet/ short link url.