Move common parts of sdl templates to additional templates (#8247)
* Move common parts of sdl templates to additional templates

SDL Validation is done in both the repos and the staging/validaiton pipelines. Today, we have a large amount of the sdl job template duplicated for the staging/validation pipelines because we need to use different sources/we don't need all the artifact downloading (and if we did, we would want to get the artifacts from a different place). This duplication leads to issues like https://github.com/dotnet/arcade/issues/8243 (among others), where versions changed in arcade, but they weren't properly migrated to dotnet-release, and the validation pipeline started failing.

This change moves the duplicated steps to a steps template, and the duplicated variables to a variables template, so we can more easily share between these repos/pipelines.

* Fix typo