Bring XHarness part of Helix SDK from main (#8258)
Backport of XHarness related functionality of Helix SDK from `main` to the `release/6.0` branch.

Fixes #8253
Related to https://github.com/dotnet/core-eng/issues/14888

## Context

As part of the **runtime-staging** pipeline in **dotnet/runtime**, we run Apple and Android simulator tests. We also run device tests in the PR build of **dotnet/arcade**.
These tests are relying on the Helix on-prem infrastructure using the XHarness submodule of Helix SDK.
As we improve the SDK in how we work with the Apple/Android infrastructure, we make the test runs more resilient and careful with how we treat the simulators/devices.

## Motivation

Last week, we have rotated Apple signing certificates in the Helix infrastructure which broke Arcade's `release/6.0` PR build (see #8253).
This backport brings the latest shape of the XHarness related part of Helix SDK so that we can keep using the infrastructure from both `main` and `release/6.0` branches.
It is probably easier to copy the whole XHarness functionality as there were no breaking changes and keep the branches in sync than to try fixing the issue only and diverging.

Bringing over all of the changes will make `release/6.0` tests more stable, will start collecting useful telemetry and will keep the infra in a better shape.