Remove graph file generation functionality (#8315)
Removes the functionality from the templates that generates graph files. These aren't really useful anymore, as BARViz subsumed this functionality a long while back. We also need to move to the 1ES pools, so this removes another place where they aren't in use.

Repos that happen to still specify should not see a break, and the additional parameters (graphFileGeneration) should just get ignored.