Remove support for v2 publishing in the templates (#8325)
Part of the 1ES pool work would require moving a bunch of legacy, deprecated jobs to the 1ES pools for completeness. I see this as an opportunity to cut off v2 publishing, which should no longer be in use by anyone. While this is not a full removal from .NET Arcade's main, it removes the crucial templating support that enables repos to get into trouble.

Benefits include:
- Simplification of the templates - The allocation time on the 1ES pools will be signficantly longer. We may need to do work in the templates to mitigate this cost. Simplifying the templates will help in this effort
- Keep people from accidentally using V2 - There have been a number of repos over the last few months found to still be using v2, and getting broken as a result of changes. V2 is no longer supported as of .NET 5. Removal should help with this.