Detect network connectivity issue in Android devices (#8263)
### Context
We often see Android devices losing their WiFi connection and going offline. In this case, we are still able to install apps via the USB cable and run most of the tests.

Some of the test apps require internet though (such as `System.Net.*.Tests`) and no connectivity usually means the tests fail. From infrastructure point of view, we have no idea that there was this problem though.

### This change

.. performs a network check from the device in case of any Android device failure and retries the work item on a different machine in case the WiFi connection is down.

We also send telemetry event so that we learn about the device without connectivity.

More details in #8232