Fix build break because of missing metdata (#8364)
* Fix build break because of missing metdata

In the previous change I didn't notice that the binplacing logic depends on metadata passing in being preserved in the items being returned. Because of that, I revert the collection change from List to HashSet and instead manually iterate through the list. The binplacing logic depends on duplicates being allowed versus the inner build project logic explicitly doesn't want duplicates to avoid unnecessary inner builds, hence adding a switch for it.

* Update Microsoft.DotNet.Build.Tasks.TargetFramework.Sdk.targets

* Update ChooseBestTargetFrameworksTask.cs