Add a timeout to WiX download
The WiX download is huge and for some reason it goes slowly on my otherwise 80 Mbps connection. The default timeout is causing an unnecessary build break that I keep hitting.

I filed this as #8301 originally (for the root cause) but since that one was closed and I hit it again and lost time on this, I'm proposing this workaround.