Remove support for NS2x runtime tfm and rename Sdk
dotnet/runtime was the only consumer that dependent on runtime specific
.NETStandard tfms. Now that we are removing them with
https://github.com/dotnet/runtime/pull/64610 we can also delete the
hacks to make them work in the TargetFramework package.

Also renaming the project from
Microsoft.DotNet.Build.Tasks.TargetFramework.Sdk to
Microsoft.DotNet.Build.Tasks.TargetFramework as doesn't need to be an
msbuild sdk anymore and consuming it as a package will reduce evaluation
time.