Add alternative casings for finding NuGet.config file. (#7549)
This allows us to find `NuGet.config` files that are named `NuGet.Config` or `nuget.config`, all of which are supported by NuGet.  In particular, `nuget.config` is actually the recommended casing for new projects on Linux so this may be increasingly common.

I didn't add an error condition when none of these are found because Arcade does not always need to work with NuGet.config, so this defers the error to when it actually causes a problem.