Define TestPlatforms.LinuxBionic, for headless Android use
We have an internal custom who needs support for devices with an Android
kernel and libc, but no Android UI (and no Android JNI). We need to be
able to treat this as Linux in most cases, but Android in a few edge
cases. This new TestPlatform definition checks for an env var which is
set on Android only, to signal "this is Linux but also Android". Then,
we can use this to skip Android-failing tests on Linux Bionic.