Support byref fields in GenAPI
Support for ref fields is being added in .NET 7 / C# 11.0.
This will impact reference assembly generation as ref fields impact
how C# consumes types that contain them. A ref field represents a
change to the metadata format and that can cause issues with tool
chains that are not updated to understand this metadata change.
A concrete example is C++/CLI which will likely error if it consumes
a ref field.

The following invariants must be preserved when not exposing ref fields:

- The containing type can never be considered unmanaged

- The type of the field, sans ref, is still important for generic
  expansion calculations