[Workloads] Allow missing workload packs
Commit f8c0d511 introduced a change that breaks MSI generation for the
Android workload.  Previously, workload packs that were declared in the
WorkloadManifest.json but not present on disk would be skipped.

The .NET 7 Android workload manifest [now declares SDK packs][0] that
have already shipped in order to support .NET 6 projects.  We do not
want to regenerate MSI installers for these packs, and we should be able
to continue shipping them in VS by keeping the .NET 6 workload authoring
in VS alongside new .NET 7 workload authoring.

A new `AllowMissingPacks` task parameter has been added to let workload
owners skip MSI and setup authoring generation for packs that are not
present when the tooling runs.

[0]: https://github.com/xamarin/xamarin-android/blob/ac5bb6b910a5d9e72f38b9df4528dd308e5000d6/src/Xamarin.Android.Build.Tasks/Microsoft.NET.Sdk.Android/WorkloadManifest.in.json#L45