Support wixpack signing in Sign.proj (#9721)
* Support wixpack signing in Sign.proj
Wixpacks were implemented for post-build signing to allow us to sign MSIs and bundles after the build was done. They are a separate zip file that contains the files necessary to reconstruct an installer. Prior to post-build signing, repositories producing installers typically signed the files going into the installer, then created the installer, then signed it. Potentially there were other stages for nested installers, dealing with bundles, etc.

There is a push to add back in-build signing. The in-build staged installer creation/signing steps have been removed from some repos. Rather than add these back, it makes more sense to enable signing installer as part of the normal signing process, which happens just before the publish stage of the build. Here the files to be signed are analyzed, unpacked, signed, and repacked.

To do this, we need a tweak to Arcade to send the path to the wix tooling to the sign task. If a repo produces no installers or wixpacks, this tooling will be unused.

While doing this, I also unified some of the wix tooling packages in use to use the same version of the package.

* Update eng/Versions.props

Co-authored-by: AraHaan <seandhunt_7@yahoo.com>

* Update src/Microsoft.DotNet.Arcade.Sdk/tools/DefaultVersions.props

Co-authored-by: AraHaan <seandhunt_7@yahoo.com>

* Revert "Update src/Microsoft.DotNet.Arcade.Sdk/tools/DefaultVersions.props"

This reverts commit 8af4cc383e051cba1927344f408aa2b757c17c89.

* Revert "Update eng/Versions.props"

This reverts commit 5f2cab65305851d9786669e6b98041546a4d0054.

* Update property name to match new package name

Co-authored-by: AraHaan <seandhunt_7@yahoo.com>