Remove (EOL 6/30/2022) Debian 9 usage. (#9848)
* Remove (EOL 6/30/2022) Debian 9 usage.

* Add FailOnMissingTargetQueue property usage so we can only fail when we have <1 queue to send to.

* Cherry pick: Separate Helix and XHarness testing (#8255)

* Separate Helix and XHarness testing

There are frequently intermittent failures in the XHarness testing that need to be retried due to timeouts or xharness flakiness. Because the XHarness tests are in the same job as the regular helix tests, we have rerun all of the testing just to rerun the xharness tests. This wastes resources and is frustrating for devs.

This change splits out the xharness tests into their own jobs so that when xharness reruns need to happen, we only have to rerun the xharness legs.

* Fix publishing logs

* And there

* Fix yaml syntax, separate even further