Get copyright info for native and r2r binaries (#10148)
* Get copyright info for native and r2r binaries
When attempting to re-enable in-build signing, runtime was throwing out an error that Microsoft.NETCore.App.Composite.r2r.dll had an empty copyright (interpreted as non-MSFT) but was being signed with a Microsoft cert. What was happening was that this r2r binary, because of the way it's created, doesn't have the assembly attributes with copyright info that would normally be present.

In this case, as well as native binaries, we should be using the file version info present in native resources to find the copyright info. This is used as a fallback for managed binaries, and always in the case of native (which didn't have copyright info before).

* Fixup tests and native resource copyright lookup

* Turn warning into high pri message

* Turn off tests that won't pass before 10293

* Differentiate 3rd party/Microsoft signing warnings