Add correct number of RemoteExecutor.Invoke overloads for Func
In https://github.com/dotnet/runtime/pull/74994 we hit an issue because a test was inadvertently using `RemoteExecutor.Invoke(Action<string, string, string, string> method, ...)` since there was no overload that takes `Func<string, string, string, string, Task>` (only one with three strings) and that means it's becoming an async void (bad!).

This PR makes sure we have a matching number of overloads that take Func with the same number of arguments as the one that takes Action.

Since there was a bit of a mismatch (one overload took five arguments), I made the overloads consistent so that everyone takes five args.