Set vsMajorVersion when provided vs version is too low (#11626)
In the "vs version provided in global.json is too low" case when initializing msbuild, we were failing to set the vsMajorVersion, resulting in the issue reported in https://github.com/dotnet/arcade/issues/11383. This change sets vsMajorVersion in that case so that the code path does not immediately fail.