Increase the timeout for publish using darc (#11695)
The timeout for the actual publishing pipeline job for publish using darc is 120 minutes, where as the pipelines waiting for the publishing job were set to the default (60 minutes). This change increases the timeout of the waiting job to 150 minutes, to account for publishing as well as any communication lag.