Begin removal of MicrosoftSourceLinkVersion (#12197)
This property is a proxy for the versions of the source link packages. Proxy properties do not generally play well with source build. Source build must explictly set this property to be equal to the version of one of the other packages (GitHub or AzureRepos.Git) in order to properly override sourcelink to the incoming version.

Fixing this will happen in three stages:
- This PR: Change internal package references to reference the dependency flow properties for each individual package (GitHub and AzureRepos.Git). Add these properties to the generated sdk versions. Retain the original property name, but set it off of the GitHub package version.
- Future PRs: Remove usages of MicrosoftSourceLinkVersion in repos. Where this property is used in repos, instead use the correct individual properties.
- Final PR: Remove MicrosoftSourceLinkVersion.

Some repos today set MicrosoftSourceLinkVersion (usually to an ancient version) but do not use it. This overrides Arcade's version. I believe most of these cases are in error, where repo owners expected dependency flow, but are not getting it. In these cases, this change will transparently upgrade them to Arcade's default.