Remove sleet and tendrils (#12301)
* Remove sleet and tendrils
I was looking to get arcade onboarded onto source-build PVP flow and ran into an issue where there are two different Nuget versions referenced in arcade. One of these is the latest one, and one is 4.4.0. The old one is apparently around because of sleet. We stopped using blob feeds in main long ago. In order to make progress, I took the opportuntity to remove sleet and a lot of tendrils of legacy publishing infras. I did not get it all.

Things that got removed:
- Old tasks that are no longer referenced (vestiges of v2 publishing)
- Sleet references and supporting types.
- Another vestige of the v2 publishing infra (PublishArtifactsInManifestV2), which was no longer referenced.

The primary goal was removing the sleet reference. As I threaded my way through doing that, I pulled out other things that happened to be in the vicinity and were no longer referenced in the following places:
- Arcade itself
- Searches on GitHub did not reveal usages (where applicable)
- Searches of dotnet-release, since it's publishing infra uses some special features.

* Delete some more files