Introduce targets to filter target frameworks (#12310)
* Introduce targets to filter target frameworks
Relies on https://github.com/dotnet/msbuild/pull/8350 to be active or not cause build errors when not enabled. This is largely still experimental and may change.
Enables the ability for an invocation of an arcade build to only build a desired set of target frameworks. The primary consumer is source-build. This allows source-build to simply remove TFMs from the build that would require SBRPs without extensive changes in project files.