Add an opt-in target to flag .NETStandard1.x dependencies
The .NETStandard1.x dependency graph is extremely big and results in dependency scan utilities flagging the brought in packages (in case of vulnerabilities). All the .NETStandard1.x packages except for Microsoft.NETCore.Platforms haven't shipped for years anymore and are considered deprecated.

Provide an opt-in target that can be enabled either per project or per repository (usually the later).