Default publishing to v3 and more cleanup (#13213)
* Default publishing to v3
Set the default publishing version to v3. This was previously done by setting this in PublishArtifactsInManifest.proj. However, this has no effect in this project now. Manifest generation is where it is critical to get the right value. This should happen in Publish.proj, PublishArtifactsInManifest, and GenerateBuildManifest. All should default to 3/Next unless specified.

This still allows for overriding publishing versions in eng/Publishing.props if it became required (e.g. v4 publishing). But makes it non-critical.

* More publishing cleanup
- Rename some old PublishingInfraVersion enum values to correctly reflect their state.
- Remove PushToBlobFeed
- Remove legacy blob feed docs