Fix some bugs with the new XunitAssert (#13510)
I found a few bugs in trying to enable https://github.com/dotnet/runtime/pull/86193.
One is that there's no way to flip Arcade to pull in the fork package instead of xunit.assert.
The second is that when both are pulled in, because the fork has a different assembly name, the
references are ambiguous and produce compile-time breaks. Matching names seems like it will get
rid of that problem, and it also has the side-effect of letting the analyzer work properly against
the fork.