Do not generate symbol packages in outer source-build (#13931)
* Improve duplicate asset error handling
Right now it's possible for a build to contain a set of assets with duplicated names. This can occur in various places:
- Merged manifest contains dupes
- Maestro build asset list contains dupes

These may not be exact dupes. You could have two assets with the same name but different versions. However, I think this scenario should generally be disallowed. A Maestro build with multiple assets with the same name would not be able to successfully flow dependencies downstream. What is chosen to flow downstream would essentially be arbitrary, which is not desirable.

So, I think we should take steps to disallow this scenario and harden against it. Fully hardening against it will require changes at multiple levels:
- Manifest merging (disallow duplicate entries even with same name + version)
- Maestro build asset upload
- Arcade publishing.

It's important to harden in multiple locations to improve error handling and avoid missing some code paths (e.g. if the merging code was updated in some branches but not others).

This PR begins this process by hardening the publishing phase. In addition, some general code cleanup and refactoring was performed.

* Do not generate symbol packages during source-build publishing
The source build intermediates simply contain nupkgs, no symbols. Generating these packages just creates a copy of the original package.

* Allow for auto-generation during inner source build